breaking things out of index.js so there's less to scroll through
Broke the .js files into index.js (stuff that happens outside of the game) and game.js (stuff that happens during the game). All the variables in the .js files are global, so if something gets initialized in one it will be available in the other. I tried to make sure that the relevant variables were mostly in the appropriate files. Remember that the variables are global and be careful when renaming things later.