Tracking mouse position
Added ability to track mouse position. For now, the data are only being output to the console before being wiped at the onset of each trial. The minimum inter-sample interval seems to be about 3 ms, with there normally being about 4 ms between samples.

Need to start saving these arrays at the end of each trial. Should also save screen width, etc. for each participant so the reaches can be normalized to % of distance to the target.