Implement search ring, timeouts for spending too long on instructions/ITI/the game, random identifier, updating instructions
Mouse pointer: set to on right now
Search ring: Now, a yellow ring with a radius equal to the distance of the mouse pointer from the center of the screen is displayed during the ITI when the participant has to get the cursor back to center. Cursor is only visible 1/8 of the way from the center to the target radius.
Timeouts: If the subject takes 1 minute to read the task instructionsm, takes 2 minutes to find the center, or spends 60 minutes in the game part of the experiment, they will be kicked out of the game.
Random identifier: A random identifier is generated for each subject. The user can change the experiment string so that there wil be a non-random sequence of characters at the beginning of the subject.id unique to the experiment.
Updating instructions: Partway through updating the instructions page. Need to make new videos for explaining the task and add them to the repo.
Other things to do: Add a field to subject specifying the json used to pull the trial parameters, remove the experimentID from the trials structure. Add fields to the trials structure to accommodate the online hand positions.