Added ability to write reach x/y/timestamp to firebase server, pick a random part of the workspace to put the targets in, update videos for instructions, cosmetic changes to webpage
Reach path: Reach paths now recorded, with up to 200 samples. Assumes that sampling cannot exceed an interval of 3-5 ms. Assumes that anything >600-1000 ms is not of interest because the reach is not in line with experiment parameters.
Randomize target location: A value between 1 and 360 is generated for the game (1 per game instance) and then added to all target angles, to randomize the target location around the workspace.
Videos: took videos while doing task
Cosmetic stuff: updates to instructions